server/requirements.txt
flask
flask-cors
flask-socketio
openai
python-decouple
python-dotenv
